Rename int <-> jun count functions
